fix: unify baseline calculation between client and server
The server-side baseline calculation was missing cumulative handling that
the client had, causing SSR-rendered charts (thumbnails, OG images) to show
incorrect values for cumulative excess charts.

Root cause: Client and server had duplicated baseline calculation code that
diverged over time. The client's calculateExcess() was updated to handle
cumulative data from the /cum endpoint, but the server's copy was not.

Solution: Extract shared calculation logic into app/lib/baseline/ module:
- core.ts: Pure calculation functions (calculateExcess, cumulativeSumFrom, etc.)
- calculateBaseline.ts: Main logic with injectable fetch dependencies
- Both client and server now use the same code, injecting their fetch impl

This prevents future divergence - any fix to calculation logic automatically
applies to both SSR and client rendering.
